requirements.txt
pyunpack        ==  0.2.1        
pysubs2         ==  0.2.4          
langdetect      ==  1.0.8       
chardet         ==  3.0.4         
iso-639         ==  0.4.5          
ffmpeg-python   ==  0.2.0  
fonttools       ==  4.12.1       
anitopy         ==  2.0.0          
pyinstaller     ==  4.3
alive-progress  ==  1.6.2
typer           ==  0.4.0
appdirs         ==  1.4.4
